Make echo be v6 aware
In fixing it up to be v6 aware, I started gdb. Being frustrated by
the client struct ot being able to be displayed, I added debug
support flags in the configure.ac.

I also added -Wall.

That triggered a complaint that assign_connector() was not
returning anything!  Which is exactly the bug that killed the
last check-in.

Signed-off-by: Tom Haynes <loghyr@gmail.com>