The list of children is placed in the inode and not the dirent.
Happy enough with the approach, not happy with the code change.

However, I know things will cahnge as it scales and goes
to a real backing store.

This change was driven by the fact that FUSE and such
employ a full path on all operations, whilst NFS and such
provide a file handle and a name for MKDIR.

As such, whilst looking up the inode is easy for NFS, looking
up the dirent is prohibative. A quick solution is to embed
the children dirents into the inode and not the parent
dirent.

The actual change is perhaps the most straight forward, allowing
the FUSE implementation to still provide testing as the
NFS layer is developed. But, it is not the cleanest.

Signed-off-by: Tom Haynes <loghyr@gmail.com>