Since this is essentially a RAM disk backed file-system, stuff some random words in the FH
So testing is a quick cycle of pjdfstest and restarting reffs.

I *think* this is confusing the Linux NFSv3 client because dirents
are cached, even across restarts. So the it gets confused
on whether a fh describes a regular file or a directory.

There are some reserved words to use to force the FHs to change, which
will cause the client to invalidate caches.

Signed-off-by: Tom Haynes <loghyr@gmail.com>